SY-132 We now select a vehicle if there is not default, so long as there are vehicles registered.
What remains is registering a vehicle if there are no registrations for the user, as well as displaying the registration number beside the charge station so that the user knows which car is charging where.

Changed text of cancel button from 'Abort' to 'Cancel' as this is more natural speech.

Changed name of last parameter of charge start to otherParams from params so as to avoid multiple params declarations.