CUSY-18 SY-132 We now show the registration number for the charging session if this is provided.
The only remaining issue is to remove currently charging vehicles from the options list for number plates as well as prompting for new number plate if the number of charging vehicles equals the number of registered vehicles.

Renamed AddFirstVehicleModalComponent to AddVehicleModalComponent as we shall use it to add vehicles which are not first if we have used up all registered vehicles.