CUSY-18 SY-132 Added a constraint on removing a vehicle if it is charging.
This is not a foolproof method. The problem is that we can only find out if a vehicle is charging after having received session data. This data we receive only on the charge page. If we have not passed through that page it means we shall not be able to know which cars are currently charging.

For regular users this should be no problem since the application always starts on the charge page. However, if a user was to use a Web browser and refresh on the settings page, it is possible to remove a vehicle even though it has an active charge session.