NL-58 Users who do not have to pay for charging got the payment-method-is-missing message.
This has now been fixed.

Also fixed is the issue whereby charge zone owners did not get a payment status message if there was an error.

This has been done in part thanks to a new property called "userPaysConsumption" which has been introduced in the back-end.

Also improved language in English version of contract satus warnings.